+++ src/Extensions/ContentControllerExtension.php
@@ -5,82 +5,102 @@ namespace Innoweb\CookieConsent\Extensions;
 use Innoweb\CookieConsent\CookieConsent;
 use Innoweb\CookieConsent\Pages\CookiePolicyPage;
 use Innoweb\CookieConsent\Pages\CookiePolicyPageController;
+use Exception;
+use SilverStripe\CMS\Controllers\ContentController;
 use SilverStripe\Control\Controller;
 use SilverStripe\Control\Director;
-use SilverStripe\Core\Extension;
 use SilverStripe\Core\Config\Config;
+use SilverStripe\Core\Extension;
+use SilverStripe\Core\Manifest\ModuleLoader;
+use SilverStripe\ORM\DataObject;
 use SilverStripe\Security\Security;
 use SilverStripe\View\Requirements;
 
+/**
+ * Class ContentControllerExtension
+ * @package Innoweb\CookieConsent
+ * @property ContentController owner
+ */
 class ContentControllerExtension extends Extension
 {
     private static $allowed_actions = [
-        'acceptCookies',
-        'revokeCookies',
+        'acceptAllCookies'
     ];
 
-    public function CookiesAccepted()
+    /**
+     * Place the necessary js and css
+     *
+     * @throws \Exception
+     */
+    public function onAfterInit()
     {
-        return CookieConsent::accepted();
+        if (!($this->owner instanceof Security) && !CookieConsent::check()) {
+            if (Config::inst()->get(CookieConsent::class, 'include_css')) {
+                Requirements::css('innoweb/silverstripe-cookie-consent:client/dist/css/cookieconsent.css');
+            }
+        }
     }
 
-    public function CookieConsent() {
-        // check controller
-        if ($this->owner instanceof Security || $this->owner instanceof CookiePolicyPageController) {
-            return false;
-        }
-        // check if cookies already accepted
-        if (CookieConsent::accepted()) {
-            return false;
-        }
-        // show cookie consent
-        if (Config::inst()->get(CookieConsent::class, 'include_javascript')) {
-            Requirements::javascript('innoweb/silverstripe-cookie-consent:client/dist/javascript/bundle.js', ['defer' => true]);
-        }
-        if (Config::inst()->get(CookieConsent::class, 'include_css')) {
-            Requirements::css('innoweb/silverstripe-cookie-consent:client/dist/css/cookie-consent.css');
-        }
-        return $this->owner->renderWith('Innoweb/CookieConsent/Includes/CookieConsent');
+    /**
+     * Method for checking cookie consent in template
+     *
+     * @param $group
+     * @return bool
+     * @throws Exception
+     */
+    public function CookieConsent($group = CookieConsent::NECESSARY)
+    {
+        return CookieConsent::check($group);
     }
-    
+
+    /**
+     * Check if we can promt for concent
+     * We're not on a Securty or Cooky policy page and have no concent set
+     *
+     * @return bool
+     */
+    public function PromptCookieConsent()
+    {
+        $controller = Controller::curr();
+        $securiy = $controller ? $controller instanceof Security : false;
+        $cookiePolicy = $controller ? $controller instanceof CookiePolicyPageController : false;
+        $hasConsent = CookieConsent::check();
+        $prompt = !$securiy && !$cookiePolicy && !$hasConsent;
+        $this->owner->extend('updatePromptCookieConsent', $prompt);
+        return $prompt;
+    }
+
+    /**
+     * Get an instance of the cookie policy page
+     *
+     * @return CookiePolicyPage|DataObject
+     */
     public function getCookiePolicyPage()
     {
-        return CookiePolicyPage::get()->first();
+        return CookiePolicyPage::instance();
     }
 
-    public function acceptCookies()
+    public function acceptAllCookies()
     {
-        CookieConsent::accept();
+        CookieConsent::grantAll();
 
         // Get the url the same as the redirect back method gets it
         $url = $this->owner->getBackURL()
             ?: $this->owner->getReturnReferer()
                 ?: Director::baseURL();
+
         $cachebust = uniqid();
-        $url = Controller::join_links(Director::absoluteURL($url), "?cache=$cachebust");
+        if (parse_url($url, PHP_URL_QUERY)) {
+            $url = Director::absoluteURL("$url&acceptCookies=$cachebust");
+        } else {
+            $url = Director::absoluteURL("$url?acceptCookies=$cachebust");
+        }
+
         $this->owner->redirect($url);
     }
 
-    public function getAcceptCookiesLink()
+    public function getAcceptAllCookiesLink()
     {
-        return Controller::join_links($this->owner->Link(), 'acceptCookies');
-    }
-
-    public function revokeCookies()
-    {
-        CookieConsent::revoke();
-        
-        // Get the url the same as the redirect back method gets it
-        $url = $this->owner->getBackURL()
-            ?: $this->owner->getReturnReferer()
-            ?: Director::baseURL();
-        $cachebust = uniqid();
-        $url = Controller::join_links(Director::absoluteURL($url), "?cache=$cachebust");
-        $this->owner->redirect($url);
-    }
-    
-    public function getRevokeCookiesLink()
-    {
-        return Controller::join_links($this->owner->Link(), 'revokeCookies');
+        return Controller::join_links($this->getOwner()->Link(), 'acceptAllCookies');
     }
 }