+++ src/Extensions/SiteConfigExtension.php
@@ -10,8 +10,11 @@ use SilverStripe\Forms\TextField;
 use SilverStripe\Forms\HTMLEditor\HTMLEditorField;
 use SilverStripe\Forms\FieldList;
 use SilverStripe\SiteConfig\SiteConfig;
-use Symbiote\GridFieldExtensions\GridFieldOrderableRows;
 
+/**
+ * Class SiteConfigExtension
+ * @package Innoweb\CookieConsent
+ */
 class SiteConfigExtension extends DataExtension
 {
     private static $db = array(
@@ -24,17 +27,17 @@ class SiteConfigExtension extends DataExtension
         'CookieConsentContent'
     );
 
+    /**
+     * @param FieldList $fields
+     * @return FieldList|void
+     */
     public function updateCMSFields(FieldList $fields)
     {
+        $fields->findOrMakeTab('Root.CookieConsent', _t(__CLASS__ . '.CookieConsent', 'Cookie Consent'));
         $fields->addFieldsToTab('Root.CookieConsent', array(
-            TextField::create('CookieConsentTitle', $this->owner->fieldLabel('CookieConsentTitle')),
-            HtmlEditorField::create('CookieConsentContent', $this->owner->fieldLabel('CookieConsentContent')),
-            GridField::create(
-                'Cookies', 
-                'Cookie Groups', 
-                CookieGroup::get(), 
-                GridFieldConfig_RecordEditor::create()->addComponents(GridFieldOrderableRows::create())
-            )
+            TextField::create('CookieConsentTitle', _t(__CLASS__ . '.CookieConsentTitle', 'Cookie Consent Title')),
+            HtmlEditorField::create('CookieConsentContent', _t(__CLASS__ . '.CookieConsentContent', 'Cookie Consent Content')),
+            GridField::create('Cookies', _t(__CLASS__ . '.Cookies', 'Cookies'), CookieGroup::get(), GridFieldConfig_RecordEditor::create())
         ));
     }
 
@@ -47,18 +50,19 @@ class SiteConfigExtension extends DataExtension
     {
         if ($config = SiteConfig::current_site_config()) {
             if (empty($config->CookieConsentTitle)) {
-                $config->CookieConsentTitle = _t(__CLASS__ . '.CookieConsentTitle', 'This website uses cookies');
+                $config->CookieConsentTitle = _t(__CLASS__ . '.DefaultCookieConsentTitle', 'This website uses cookies');
             }
 
             if (empty($config->CookieConsentContent)) {
-                $config->CookieConsentContent = _t(__CLASS__ . '.CookieConsentContent', '<p>This website or its third-party tools use cookies, which are necessary for its functioning and required to achieve the purposes illustrated in the cookie policy.<br>You accept the use of cookies by closing or dismissing this notice, by clicking a link or button or by continuing to browse otherwise.</p>');
+                $config->CookieConsentContent = _t(__CLASS__ . '.DefaultCookieConsentContent', '<p>We use cookies to personalise content, to provide social media features and to analyse our traffic. We also share information about your use of our site with our social media and analytics partners who may combine it with other information that you’ve provided to them or that they’ve collected from your use of their services. You consent to our cookies if you continue to use our website.</p>');
             }
 
             $config->write();
         }
     }
-    
-    public function getCookieGroups() {
+
+    public function getCookieGroups()
+    {
         return CookieGroup::get();
     }
 }