+++ changelog.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.0.0]
+
+rebuild most of the module using https://github.com/TheBnl/silverstripe-cookie-consent for inspiration:
+* list individual cookies on cookie policy page
+* add form to accept different cookie categories
+* add translations
+
 ## [1.0.1]
 
 * update dependencies